add array support for debug impl
